pyproject.toml
[tool.poetry]
name = "authproxy"
version = "0.1.0"
description = ""
authors = ["4thel00z <4thel00z@gmail.com>"]
packages = [
	{ include = "authproxy"},
	{ include = "libauthproxy"}
]

[tool.poetry.dependencies]
python = "^3.10"
edgedb = "1.5.0"
fastapi = "0.98.0"
ujson = "^5.7.0"
uvicorn = "0.22.0"
python-dotenv = "1.0.0"
anyio = "3.7.0"
bcrypt = "4.0.1"
cffi = "1.15.1"
click = "8.1.3"
cryptography = "41.0.1"
ecdsa = "0.18.0"
exceptiongroup = "1.1.1"
h11 = "0.14.0"
idna = "3.4"
passlib = "1.7.4"
pyasn1 = "0.5.0"
pycparser = "2.21"
pydantic = "1.10.9"
python-jose = "3.3.0"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.0"
starlette = "0.27.0"
typing-extensions = "4.6.3"
python-multipart = "^0.0.6"

[tool.poetry.dev-dependencies]
black = "^23.3.0"
httpie = "^3.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
